+++ app/src/main/java/com/example/listary/view/Pantry/PantryActivity.java
@@ -16,6 +16,7 @@ import android.widget.EditText;
 import android.widget.Toast;
 
 import com.example.listary.R;
+import com.example.listary.controllers.PantryController;
 import com.example.listary.model.Pastry;
 import com.example.listary.view.createProduct.SearchProductActivity;
 import com.example.listary.view.historic.HistoricActivity;
@@ -41,6 +42,8 @@ public class PantryActivity extends AppCompatActivity {
     private FirebaseFirestore db = FirebaseFirestore.getInstance();
     FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
 
+    private PantryController pantryController = new PantryController();
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -51,36 +54,14 @@ public class PantryActivity extends AppCompatActivity {
         edPantry = findViewById(R.id.edPantry);
 
         btnSavePantry = findViewById(R.id.btnSavePentry);
-        show();
+        pantryController.show(edPantry, uid);
         btnSavePantry.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                post();
+                pantryController.post(edPantry, uid);
             }
         });
     }
-
-    private void show() {
-        DocumentReference docRef = db.collection("data").document(uid).collection("pantry").document("FQpG5QWFiJ4xStsiDING");
-        docRef.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
-            @Override
-            public void onComplete(@NonNull Task<DocumentSnapshot> task) {
-                if (task.isSuccessful()) {
-                    DocumentSnapshot document = task.getResult();
-                    if (document.exists()) {
-                        String banco = document.getString("description");
-                        edPantry.setText(banco);
-                        edPantry.setSelection(edPantry.getText().length());
-                    } else {
-                        Log.d("n", "No such document");
-                    }
-                } else {
-                    Log.d("erro", "get failed with ", task.getException());
-                }
-            }
-        });
-    }
-
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
@@ -100,56 +81,34 @@ public class PantryActivity extends AppCompatActivity {
             case R.id.menuListary:
                 intent = new Intent(this, MenuListaryActivity.class);
                 startActivity(intent);
+                finish();
                 return true;
 
             case R.id.novaLista:
                 intent = new Intent(this, NewListActivity.class);
                 startActivity(intent);
+                finish();
                 return true;
 
             case R.id.consultarListas:
                 intent = new Intent(this, SearchProductActivity.class);
                 startActivity(intent);
+                finish();
                 return true;
 
             case R.id.historic:
                 intent = new Intent(this, HistoricActivity.class);
                 startActivity(intent);
+                finish();
                 return true;
 
             case R.id.logOut:
                 FirebaseAuth.getInstance().signOut();
                 startActivity(new Intent(this, Login.class));
+                finish();
             default:
                 return true;
         }
-    }
-
-    private void post() {
-        String description = edPantry.getText().toString();
-        Pastry pantry = new Pastry(
-            description
-    );
-
-        if (description.isEmpty()){
-            Toast.makeText(this, "Campo não pode estar vazio", Toast.LENGTH_LONG).show();
-        } else{
-            DocumentReference documentReference = db.collection("data").document(uid).collection("pantry").document("FQpG5QWFiJ4xStsiDING");
-            documentReference.set(pantry)
-                    .addOnSuccessListener(new OnSuccessListener<Void>() {
-                        @Override
-                        public void onSuccess(Void unused) {
-                            Toast.makeText(PantryActivity.this, "Sucesso", Toast.LENGTH_SHORT).show();
-                        }
-                    })
-                    .addOnFailureListener(new OnFailureListener() {
-                        @Override
-                        public void onFailure(@NonNull Exception e) {
-                            Toast.makeText(PantryActivity.this, "Erro", Toast.LENGTH_SHORT).show();
-                        }
-                    });
-        }
-
     }
 }
